phy: samsung: exynos5250-usb2: Fix USB2 OTG mode
The following changes are needed for USB2 of Exynos 5250 devices to work
in host, device and OTG modes.

In exynos5250_isol(), the PMU offset must always be
EXYNOS5250_USB_ISOL_HOST_OFFSET.

In exynos5250_power_on(), initialization of HOST and HSIC is needed for
all cases of inst->cfg->id (that is: EXYNOS5250_HOST, EXYNOS5250_HSIC0,
EXYNOS5250_HSIC1 and EXYNOS5250_DEVICE).
Hence, the associated code can be moved outside the switch(inst->cfg-id)
statement to avoid repeating the exact same code twice.
Also, for host-only operation, a call to regmap_update_bits() for
switching to host mode, and a call to writel() for applying OTG
configuration were missing.

Signed-off-by: Alexandre Marquet <tb@a-marquet.fr>