power: supply: smb347-charger: proper mains/usb supply_prop returns
In the current driver, when smb347_usb is active and smb347_mains
inactive (and vice-versa), then querying smb347_mains
POWER_SUPPLY_PROP_CONSTANT_CHARGE_VOLTAGE and
POWER_SUPPLY_PROP_CONSTANT_CHARGE_CURRENT returns the values applied to
smb347_usb.

This commit fix this issue, by returning -ENODATA if the queried supply
is not active.

This commits also move the checks for the active supply from
smb347_get_property_locked() to smb347_charging_status() in the case of
querying the POWER_SUPPLY_CHARGE_TYPE_PROPERTY, for better consistency
with the rest of the properties.

Signed-off-by: Alexandre Marquet <tb@a-marquet.fr>