iio: light: bh1750: implement reset sequence
According to the datasheets of chips supported by the bh1750 driver, DVI
pin should be pulled low during more than 1µs after powering the chip,
in order to reset it.

This commit adds support for implementing this reset sequence with GPIOs.
It is made optional, since such a reset sequence can also be implemented
in hardware.

Signed-off-by: Alexandre Marquet <tb@a-marquet.fr>