extcon: max77693: Clear extcon status before detection
The way this is done currently is very subtly broken when switching
from one cable type to another - there is a "disconnected" check
which seems to work for the charger check, but is broken for ADC.
Even if it wasn't, the driver never clears the previous cable type,
instead disconnecting the currently detected one. This is a whole
mess that would require a more extensive investigation to figure
out, but for now, this lets us swap between OTG and charger cables
without it locking up on the OTG cable.

Signed-off-by: Artur Weber <aweber.kernel@gmail.com>