ASoC: mfd: wm8994: Allow to configure DRC from dts
Similar to wm8904. DRC (Dynamic Range Control) configuration is used to
process input and/or output audio signals depending on the use case
(e.g.: speaker, audio jack).

The set of names and configurations for DRC are specified by system
integrators. The names are exposed directly to userspace as options that
ca be selected at runtime.

Adding the DRC to the DT eases the transition from pdata. The parameters
filled in here are almost certainly specific tuning for the hardware.

Signed-off-by: Alexandre Marquet <tb@a-marquet.fr>