extcon: max77693: Implement av-cable-as-otg
Some chips report OTG cables as audio video cables. An example for this is the
Maxim MAX77888. This implements a device tree option to bring OTG functions to
devices with those chips. The setup needed for OTG is moved into a separate
function that's also called if the OTG cable is reported as an audio video
cable.

Signed-off-by: Eisenbahnfan <hendrik.noack@protonmail.com>